refactor(ffi): promote ProtocolRepoVariant to scp-ffi-common
The NAPI and UniFFI bridges each carried an identical
`ProtocolRepoVariant { InMemory(…), Sqlite(…) }` enum plus identical
`event_log_provider()` dispatch. Two identical match statements with
no type-safety benefit: each bridge already has its own concrete
instance struct, so the enum lives on a typed field with no ambiguity.

Move the enum + dispatch into `scp_ffi_common::bridge_runtime`. NAPI
and UniFFI re-export it via `pub use` so existing
`crate::runtime::ProtocolRepoVariant` references across each bridge
keep compiling without a mass rename.

ADR-048 §2 now documents the "shared-variant types for storage-backed
repositories" exemption: the rule "per-bridge concrete structs, no
shared type-erased slots" holds, but closed protocol-level enums
whose variants trace to a configuration type (`StorageConfig`) are
the one approved exception. The exemption is narrow — it does NOT
re-admit `Box<dyn Any>`, runtime downcasts, or open-ended trait
objects. Those remain rejected per Phase 4a.

PyO3's `StorageProvider` is NOT yet folded into `ProtocolRepoVariant`:
it backs the `Storage`-trait surface used by identity/event_log/mcp
(raw bytes access), which the repo layer does not expose. Migration
to the common variant would require switching PyO3's in-memory
backend from `scp_platform::testing::InMemoryStorage` to
`BridgeInMemoryStorage` and reworking those callers. Tracked as
follow-up; the split-brain risk the variant guards against is
already patched for PyO3 by the previous fix-3 commit (None arm now
returns `SCP-VALID-7005` instead of a silent in-memory fallback).

PR #1690 review — architecture should-fix (fix 5 + fix 6 ADR paragraph).
