+++ .docs/adrs/ADR-048-scp-multi-instance.md
@@ -49,6 +49,8 @@ The class is named after the protocol, not after internal plumbing. This matches
 
 Every shared helper in `scp-ffi-common` operates on `&dyn BridgeInstanceCore`. Per-bridge callers pass their concrete instance. The four `Box<dyn Any>` slots introduced in Phase 4a are removed. Type safety is compile-time; there are no runtime downcasts. This satisfies the CLAUDE.md rule "enforce mechanically — type system over documentation."
 
+**Approved exemption — shared-variant types for storage-backed repositories.** The rule "per-bridge concrete structs, no shared type-erased slots" has one deliberate carve-out: types that enumerate a **closed, protocol-level** set of storage backends. `ProtocolRepoVariant { InMemory(…), Sqlite(…) }` is the canonical case. It is not bridge plumbing — the variants are dictated by `StorageConfig` (itself protocol-shaped: in-memory vs. persistent SQLCipher), the `Storage` trait is `scp_platform`-level, and every non-WASM bridge dispatches identically over the same arms. Duplicating the enum into `NapiProtocolRepoVariant` + `UniffiProtocolRepoVariant` + `PyProtocolRepoVariant` produces three identical match statements with no extra type safety — each bridge already owns its own concrete instance type, so the enum lives on a per-bridge field without ambiguity. The exemption is **narrow**: it applies only to closed-set enums whose variants trace to a protocol-level configuration type (today: `StorageConfig`). It does not re-admit `Box<dyn Any>`, runtime downcasts, or open-ended trait objects — those remain rejected per Phase 4a. `ProtocolRepoVariant` has been promoted into `scp-ffi-common` so the three bridges share the single definition; the carve-out is documented here so future maintainers do not re-litigate the decision.
+
 ### 3. Default-instance façade remains as a sunset scaffold
 
 `DEFAULT_BRIDGE_INSTANCE: OnceLock<Arc<{Py,Napi,Uniffi}BridgeInstance>>` (renamed from `BRIDGE_INSTANCE`; the concrete type varies per bridge — `PyBridgeInstance` in PyO3, `NapiBridgeInstance` in NAPI, `UniffiBridgeInstance` in UniFFI) stays in each bridge for one deprecation window. Existing free-function exports (`py_context_create`, `napi context_create`, UniFFI `context_create`) continue to work by forwarding to `SCP::default()`. Each forward emits a one-time deprecation warning per function name:
