+++ crates/scp-ffi/uniffi/src/runtime.rs
@@ -44,15 +44,12 @@ use std::time::Duration;
 use dashmap::DashMap;
 use scp_core::context::builder::{ContextCryptoProvider, ContextEventLogProvider};
 use scp_core::context::manager::ContextManager;
-use scp_core::context::providers::MerkleEventLogProvider;
 use scp_core::crypto::ucan::nonce::NonceTracker;
 use scp_core::crypto::ucan::revoke::RevocationList;
 use scp_core::store::ProtocolRepository;
-use scp_core::store::context::ProtocolRepositoryEventLogBridge;
 use scp_event_log::EventLog;
 use scp_identity::cache::SystemClock;
 use scp_platform::encrypting_adapter::EncryptingAdapter;
-use scp_platform::sqlite::SqliteStorage;
 
 // ---------------------------------------------------------------------------
 // UniffiBridgeInstance — per-bridge concrete bridge instance (#1549 Phase 4 PR 1)
@@ -122,54 +119,11 @@ impl std::fmt::Display for StorageInitError {
 
 impl std::error::Error for StorageInitError {}
 
-/// Protocol repository variant: an `Arc<ProtocolRepository<_>>` whose inner
-/// `Storage` matches the bridge's configured persistence backend.
-///
-/// Before this variant existed, `UniffiBridgeInstance::protocol_repository`
-/// was always `Arc<ProtocolRepository<EncryptingAdapter<BridgeInMemoryStorage>>>`,
-/// even when the bridge was constructed with [`StorageConfig::Sqlite`]. That
-/// meant the Merkle event log — which uses the protocol repository as its
-/// backing `EventLogPersistence` — silently ran against an ephemeral
-/// in-memory store, while context snapshots correctly landed in `SQLite`. On
-/// restart the event log would be empty even though the rest of the state
-/// survived, producing a split-brain the caller had no way to detect.
-///
-/// The enum dispatches the event log bridge and the trust bridge onto the
-/// real backing store for each variant, so `SCP({storage: sqlite})` now
-/// persists *both* snapshots and Merkle event log entries to the same
-/// `SQLCipher` database.
-pub enum ProtocolRepoVariant {
-    /// Encrypted in-memory repository. Event log and trust aggregation are
-    /// backed by an `EncryptingAdapter<BridgeInMemoryStorage>` with a random
-    /// per-instance AES-256-GCM key. Data is lost when the instance drops.
-    InMemory(Arc<ProtocolRepository<EncryptingAdapter<BridgeInMemoryStorage>>>),
-    /// SQLCipher-backed repository. Event log and trust aggregation share the
-    /// same `Arc<SqliteStorage>` that backs `CoreFields::persistence`, so
-    /// context snapshots, trust attestations, and event log entries all
-    /// survive restart and share a single `SQLCipher` connection.
-    Sqlite(Arc<ProtocolRepository<Arc<SqliteStorage>>>),
-}
-
-impl ProtocolRepoVariant {
-    /// Constructs a [`ContextEventLogProvider`] backed by this repository.
-    ///
-    /// The bridge is retained by `Arc` inside [`MerkleEventLogProvider`], so
-    /// subsequent `append` calls persist entries through the backing store
-    /// that was configured at instance-construction time.
-    #[must_use]
-    pub fn event_log_provider(&self) -> Box<dyn ContextEventLogProvider> {
-        match self {
-            Self::InMemory(repo) => {
-                let bridge = ProtocolRepositoryEventLogBridge::new(Arc::clone(repo));
-                Box::new(MerkleEventLogProvider::with_persistence(Arc::new(bridge)))
-            }
-            Self::Sqlite(repo) => {
-                let bridge = ProtocolRepositoryEventLogBridge::new(Arc::clone(repo));
-                Box::new(MerkleEventLogProvider::with_persistence(Arc::new(bridge)))
-            }
-        }
-    }
-}
+// `ProtocolRepoVariant` lives in `scp-ffi-common::bridge_runtime`. Re-exported
+// here so existing `crate::runtime::ProtocolRepoVariant` references across
+// the UniFFI bridge keep compiling without mass rename. See ADR-048 §2 for
+// the "shared-variant types for storage-backed repositories" exemption.
+pub use scp_ffi_common::bridge_runtime::ProtocolRepoVariant;
 
 /// `UniFFI`-specific concrete bridge instance.
 ///
